# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HealthDocs - Healthcare Credential Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves credential document files. Requires authentication; access is
        limited to the file owner and manager roles.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Not allowed to access this object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /reports/compliance:
    get:
      operationId: getComplianceReport
      tags: [reports]
      summary: Generate compliance report
      parameters:
        - name: facilityId
          in: query
          description: Optional facility selector for system administrators.
          schema:
            type: number
        - name: departmentId
          in: query
          schema:
            type: ["number", "null"]
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: Compliance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceReport"
components:
  schemas:
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: number
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ComplianceReport:
      type: object
      required: [generatedAt, facilityName, overallComplianceRate, departments]
      properties:
        generatedAt:
          type: string
        facilityName:
          type: string
        overallComplianceRate:
          type: number
        departments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentComplianceDetail"
    DepartmentComplianceDetail:
      type: object
      required: [departmentName, employeeCount, complianceRate, employees]
      properties:
        departmentName:
          type: string
        employeeCount:
          type: number
        complianceRate:
          type: number
        expiredCount:
          type: number
        expiringCount:
          type: number
        employees:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeComplianceSummary"
    EmployeeComplianceSummary:
      type: object
      required: [name, nameAr, employeeNumber, complianceRate, expiredCount, expiringCount, missingCount]
      properties:
        name:
          type: string
        nameAr:
          type: string
        employeeNumber:
          type: string
        complianceRate:
          type: number
        expiredCount:
          type: number
        expiringCount:
          type: number
        missingCount:
          type: number